Placed individual widgets in separate dart files.
