+++ lib/config.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/foundation.dart';
+import 'dart:io';
+import 'package:flutter/material.dart';
+
+Socket sock;
